+++ src/Components/Providers/EncryptedStorageProvider/ApplicationSecurityProvider.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react"
 import { MMKV } from "react-native-mmkv"
-import { debug, error, HexUtils, PlatformUtils, warn } from "~Utils"
+import { BiometricsUtils, debug, error, HexUtils, info } from "~Utils"
 import { BiometricState, SecurityLevelType, WALLET_STATUS } from "~Model"
 import {
     PreviousInstallation,
@@ -9,15 +9,11 @@ import {
     StorageEncryptionKeyHelper,
     WalletEncryptionKeyHelper,
 } from "~Components/Providers"
-import Onboarding from "~Components/Providers/EncryptedStorageProvider/Helpers/Onboarding"
+
 import { useAppState, useBiometrics } from "~Hooks"
 import { StandaloneAppBlockedScreen, StandaloneLockScreen } from "~Screens"
 import { AnimatedSplashScreen } from "../../../AnimatedSplashScreen"
-import { GlobalEventEmitter, LOCK_APP_EVENT } from "~Events"
-import {
-    CANCEL_AUTH_MESSAGE_ANDROID,
-    CANCEL_AUTH_MESSAGE_IOS,
-} from "~Constants/Constants/Errors/ErrorsConstants"
+import Onboarding from "./Helpers/Onboarding"
 
 const UserEncryptedStorage = new MMKV({
     id: "user_encrypted_storage",
@@ -59,6 +55,7 @@ type IApplicationSecurity = {
     isAppReady: boolean
     setIsAppReady: (isReady: boolean) => void
     lockApplication: () => void
+    triggerAutoLock: () => void
 }
 
 const ApplicationSecurityContext = React.createContext<
@@ -69,6 +66,7 @@ type ApplicationSecurityContextProviderProps = { children: React.ReactNode }
 export const ApplicationSecurityProvider = ({
     children,
 }: ApplicationSecurityContextProviderProps) => {
+    const [autoLock, setAutoLock] = useState(false)
     const [walletStatus, setWalletStatus] = useState<WALLET_STATUS>(
         WALLET_STATUS.NOT_INITIALISED,
     )
@@ -87,6 +85,10 @@ export const ApplicationSecurityProvider = ({
     const biometrics = useBiometrics()
 
     const onboardingKey = useMemo(() => HexUtils.generateRandom(256), [])
+
+    const triggerAutoLock = () => {
+        setAutoLock(true)
+    }
 
     const updateSecurityType = useCallback(
         (type: SecurityLevelType | undefined) => {
@@ -147,26 +149,11 @@ export const ApplicationSecurityProvider = ({
                 backUpKeys?.storage ??
                 (await StorageEncryptionKeyHelper.get(pinCode))
         } catch (e) {
-            if (PlatformUtils.isIOS()) {
-                if (
-                    e instanceof Error &&
-                    e.message.includes(CANCEL_AUTH_MESSAGE_IOS)
-                ) {
-                    return unlock()
-                } else {
-                    throw e
-                }
-            }
-            if (PlatformUtils.isAndroid()) {
-                if (
-                    e instanceof Error &&
-                    (e.message.includes("Cancel") ||
-                        e.message.includes(CANCEL_AUTH_MESSAGE_ANDROID))
-                ) {
-                    return unlock()
-                } else {
-                    throw e
-                }
+            // handle cases when  the user cancels the biometric prompt and the keychain returns it as an error
+            if (BiometricsUtils.BiometricErrors.isBiometricCanceled(e)) {
+                return unlock()
+            } else {
+                throw e
             }
         }
 
@@ -233,7 +220,7 @@ export const ApplicationSecurityProvider = ({
             const encryptedStorageKeys = UserEncryptedStorage.getAllKeys()
 
             if (encryptedStorageKeys.length === 0) {
-                warn("No keys found in encrypted storage, user is onboarding")
+                info("No keys found in encrypted storage, user is onboarding")
 
                 await PreviousInstallation.clearOldStorage()
 
@@ -247,7 +234,6 @@ export const ApplicationSecurityProvider = ({
                 })
                 setWalletStatus(WALLET_STATUS.FIRST_TIME_ACCESS)
             } else {
-                warn("Keys found in encrypted storage")
                 setWalletStatus(WALLET_STATUS.LOCKED)
                 await setUpEncryptionKeys(_biometrics)
             }
@@ -320,16 +306,12 @@ export const ApplicationSecurityProvider = ({
     )
 
     const lockApplication = useCallback(() => {
-        if (
-            walletStatus !== WALLET_STATUS.UNLOCKED ||
-            currentState === "active"
-        )
-            return
+        if (walletStatus !== WALLET_STATUS.UNLOCKED) return
         setWalletStatus(WALLET_STATUS.NOT_INITIALISED)
         setReduxStorage(undefined)
         setImageStorage(undefined)
         setMetadataStorage(undefined)
-    }, [walletStatus, currentState])
+    }, [walletStatus])
 
     /**
      * Initialise the app
@@ -350,12 +332,23 @@ export const ApplicationSecurityProvider = ({
         }
     }, [walletStatus, currentState, biometrics, intialiseApp])
 
+    // Handle auto lock request
     useEffect(() => {
-        GlobalEventEmitter.on(LOCK_APP_EVENT, lockApplication)
-        return () => {
-            GlobalEventEmitter.removeListener(LOCK_APP_EVENT, lockApplication)
+        if (autoLock) {
+            try {
+                if (
+                    walletStatus === WALLET_STATUS.UNLOCKED &&
+                    currentState !== "active"
+                ) {
+                    lockApplication()
+                }
+            } catch (e) {
+                error("Error locking application", e)
+            } finally {
+                setAutoLock(false)
+            }
         }
-    }, [lockApplication])
+    }, [autoLock, setAutoLock, walletStatus, currentState, lockApplication])
 
     const value: IApplicationSecurity | undefined = useMemo(() => {
         if (!reduxStorage || walletStatus === WALLET_STATUS.NOT_INITIALISED)
@@ -374,6 +367,7 @@ export const ApplicationSecurityProvider = ({
             updateSecurityMethod,
             securityType,
             lockApplication,
+            triggerAutoLock,
         }
     }, [
         reduxStorage,