+++ src/Events/GlobalEventEmitter/GlobalEventEmitter.ts
@@ -1,7 +1,0 @@
-import { EventEmitter } from "events"
-
-export const LOCK_APP_EVENT = "lock-application"
-
-const eventEmitter = new EventEmitter()
-
-export default eventEmitter