+++ src/Events/GlobalEventEmitter/index.ts
@@ -1,3 +1,0 @@
-import GlobalEventEmitter from "./GlobalEventEmitter"
-export * from "./GlobalEventEmitter"
-export { GlobalEventEmitter }